chore: addressed PR feedback
