+++ lib/transaction/trace/index.js
@@ -11,7 +11,6 @@ const TraceSegment = require('./segment')
 const { Attributes, MAXIMUM_CUSTOM_ATTRIBUTES } = require('../../attributes')
 const logger = require('../../logger').child({ component: 'trace' })
 const { DESTINATIONS } = require('../../config/attribute-filter')
-const { mergeRanges } = require('./exclusive-time-calculator')
 const FROM_MILLIS = 1e-3
 const ATTRIBUTE_SCOPE = Attributes.SCOPE_TRANSACTION
 const REQUEST_URI_KEY = 'request.uri'
@@ -252,14 +251,75 @@ Trace.prototype._collectSegments = function _collectSegments() {
   return segmentTree
 }
 
-Trace.prototype._updateExclusiveDurationForSegment = function _updateExclusiveDurationForSegment() {
+/**
+ * Merges two sorted arrays of time ranges into a single sorted array of non-overlapping ranges.
+ *
+ * Each range is a two-element array `[start, end]` representing a time interval.
+ * When ranges overlap, they are combined into a single range spanning the full extent.
+ *
+ * **Note:** This function is destructive — it mutates end times of elements in the input
+ * arrays when merging overlapping intervals. Do not reuse the input arrays after calling this.
+ *
+ * @private
+ * @param {Array} first - A sorted array of `[start, end]` time ranges.
+ * @param {Array} second - A sorted array of `[start, end]` time ranges.
+ * @returns {Array} A new sorted array of merged, non-overlapping time ranges.
+ */
+Trace.prototype._mergeRanges = function _mergeRanges(first, second) {
+  if (!first.length) {
+    return second
+  }
+
+  if (!second.length) {
+    return first
+  }
+
+  const res = []
+  let resIdx = 0
+  let firstIdx = 0
+  let secondIdx = 0
+  // N.B. this is destructive, it will be updating the end times for range arrays in
+  // the input arrays.  If we need to reuse these arrays for anything, this behavior
+  // must be changed.
+  let currInterval =
+    first[firstIdx][0] < second[secondIdx][0] ? first[firstIdx++] : second[secondIdx++]
+
+  while (firstIdx < first.length && secondIdx < second.length) {
+    const next = first[firstIdx][0] < second[secondIdx][0] ? first[firstIdx++] : second[secondIdx++]
+    if (next[0] <= currInterval[1]) {
+      // if the segment overlaps, update the end of the current merged segment
+      currInterval[1] = Math.max(next[1], currInterval[1])
+    } else {
+      // if there is no overlap, start a new merging segment and push the old one
+      res[resIdx++] = currInterval
+      currInterval = next
+    }
+  }
+
+  const firstIsRemainder = firstIdx !== first.length
+  const remainder = firstIsRemainder ? first : second
+  let remainderIdx = firstIsRemainder ? firstIdx : secondIdx
+
+  // merge the segments overlapping with the current interval
+  while (remainder[remainderIdx] && remainder[remainderIdx][0] <= currInterval[1]) {
+    currInterval[1] = Math.max(remainder[remainderIdx++][1], currInterval[1])
+  }
+
+  res[resIdx++] = currInterval
+
+  // append the remaining non-overlapping ranges
+  for (; remainderIdx < remainder.length; ++remainderIdx) {
+    res[resIdx++] = remainder[remainderIdx]
+  }
+
+  return res
 }
 
 /**
  * Computes total time and exclusive durations for all segments in a single
  * post-order traversal of the segment tree.  Avoids repeated tree lookups by
  * working directly with SegmentTree nodes and propagating merged time ranges
- * up the tree, mirroring the ExclusiveCalculator algorithm.
+ * up the tree.
  *
  * @private
  * @returns {number} sum of exclusive durations for all non-root segments
@@ -279,7 +339,7 @@ Trace.prototype._computeTotalTime = function _computeTotalTime() {
     // Merge all direct children's propagated ranges
     let childPairs = []
     for (const child of children) {
-      childPairs = mergeRanges(childPairs, child._ranges)
+      childPairs = this._mergeRanges(childPairs, child._ranges)
       child._ranges = null
     }
 
@@ -292,7 +352,7 @@ Trace.prototype._computeTotalTime = function _computeTotalTime() {
     }
 
     // Propagate this segment's full range merged with child ranges to its parent
-    node._ranges = mergeRanges(childPairs, [segment.timer.toRange()])
+    node._ranges = this._mergeRanges(childPairs, [segment.timer.toRange()])
 
     if (node !== rootNode) {
       total += segment.getExclusiveDurationInMillis()
