+++ lib/transaction/trace/segment.js
@@ -221,7 +221,7 @@ function _setExclusiveDurationInMillis(duration) {
 }
 
 /**
- * Calculated exclusive duration for a segment which is segment duration - overlap with merged child ranges
+ * Calculate, and cache, exclusive duration for a segment which is segment duration - overlap with merged child ranges
  * @param {Array} childPairs list of non-overlapping start/end times of the segment's children [start, end]
  */
 TraceSegment.prototype.calculateExclusiveDuration = function calculateExclusiveDuration(childPairs) {
